requirements.txt
torch~=1.3.0
torchvision~=0.4.1
django==3.0.6
scipy==1.2.1
opencv-python~=4.0.0.21
scikit-image==0.17.1
scikit-learn==0.21.2
torchnet~=0.0.4
pillow~=6.2.0
tensorboardx~=2.0
numpy~=1.16.2
sklearn~=0.0